pyproject.toml
[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rag-service"
version = "0.1.0"
description = "RAG Pipeline"
authors = ["liuhaochen <liu_hao_chen@hotmail.com>"]
readme = "README.md"
packages = [{ include = "rag_service", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.6", extras = ["standard"] }
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"
sqlalchemy = { version = "^2.0.36", extras = ["asyncio"] }
asyncpg = "^0.30.0"
pgvector = "^0.3.6"
tenacity = "^9.0.0"
httpx = "^0.28.1"
orjson = "^3.10.11"
alembic = "^1.11.1"
sqlmodel = "^0.0.22"
llama-index-core = "^0.14.12"
llama-index-embeddings-google-genai = "^0.3.1"
llama-index-vector-stores-postgres = "^0.7.2"
langchain-text-splitters = "^1.1.0"

[tool.poetry.group.notebook]
optional = true

[tool.poetry.group.notebook.dependencies]
jupyterlab = "^4.3.0"
ipykernel = "^6.29.5"
numpy = "^2.1.3"
pandas = "^2.2.3"
matplotlib = "^3.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
ruff = "^0.7.2"
black = "^24.10.0"
mypy = "^1.13.0"
pre-commit = "^4.0.1"
types-requests = "^2.32.0.20241016"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
